+++ services/src/main/java/org/keycloak/services/clientregistration/ClientRegistrationAuth.java
@@ -134,7 +134,7 @@ public class ClientRegistrationAuth {
         RegistrationAuth registrationAuth = RegistrationAuth.ANONYMOUS;
 
         if (isBearerToken()) {
-            checkClientProtocol();
+           // checkClientProtocol();
 
             if (hasRole(AdminRoles.MANAGE_CLIENTS, AdminRoles.CREATE_CLIENT)) {
                 registrationAuth = RegistrationAuth.AUTHENTICATED;