+++ CHANGELOG.md
@@ -18,6 +18,9 @@ Our Keycloak version is working well with PostgreSQL database. For using other S
 ### Changed
 - SAML SP Descriptor keys for use both for encrypt and signing
 
+## Fixed
+- Client registration service do not check client protocol for Bearer token  [RCIAM-1186](https://jira.argo.grnet.gr/browse/RCIAM-1186)
+
 ## [18.0.1-2.7] - 2022-11-18
 
 ### Fixed